// legacy-src/database/repositories/SnowflakeStructure/types.ts
export interface SnowflakeStructureFile {
  id: string
  central_idia: string | null
  expansion_to_paragraph_phrase_1: string | null
  expansion_to_paragraph_phrase_2: string | null
  expansion_to_paragraph_phrase_3: string | null
  expansion_to_paragraph_phrase_4: string | null
  expansion_to_paragraph_phrase_5: string | null
  expansion_to_page_paragraph_1: string | null
  expansion_to_page_paragraph_2: string | null
  expansion_to_page_paragraph_3: string | null
  expansion_to_page_paragraph_4: string | null
  expansion_to_page_paragraph_5: string | null
  interweaving_persons_and_expansion: string | null
  created_at: Date
  updated_at: Date
}